feat(certificate): consolidate certification utilities and clean obsolete TODOs
  - Move generateEmbedCodes from certificateUtils to certificationConstants
  - Update CertificateBadgeEmbedModal to use unified certification API
  - Remove deprecated certificateUtils.js (getCertificateLevelLabel, getBadgeCdnUrl, getCertificateBadgeComponent)
  - Delete obsolete TODOs for non-existent CertificationBronze/Argent/Or/Diamant components
  - Keep only relevant TODO for CDN URL update after master merge
